# chat-service/docs/swagger.yaml
basePath: /
definitions:
  dto.ChatResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  dto.CreateChatRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  dto.CreateChatResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  dto.CreateMessageRequest:
    properties:
      text:
        type: string
    required:
    - text
    type: object
  dto.CreateMessageResponse:
    properties:
      chat_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      sender_id:
        type: integer
      text:
        type: string
    type: object
  dto.ListChatsResponse:
    properties:
      chats:
        items:
          $ref: '#/definitions/dto.ChatResponse'
        type: array
    type: object
  dto.ListMessagesResponse:
    properties:
      messages:
        items:
          $ref: '#/definitions/dto.MessageResponse'
        type: array
    type: object
  dto.MessageResponse:
    properties:
      chat_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      sender_id:
        type: integer
      text:
        type: string
    type: object
host: localhost:8081
info:
  contact: {}
  description: Chat service for messenger
  title: Chat Service API
  version: "1.0"
paths:
  /chats:
    get:
      description: Get a list of all chats
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListChatsResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all chats
      tags:
      - chats
    post:
      consumes:
      - application/json
      description: Create a new chat with the given name
      parameters:
      - description: Chat name
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateChatRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CreateChatResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new chat
      tags:
      - chats
  /chats/{id}:
    get:
      description: Get a chat by its ID
      parameters:
      - description: Chat ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ChatResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get chat by ID
      tags:
      - chats
  /chats/{id}/messages:
    get:
      description: Get all messages from a chat
      parameters:
      - description: Chat ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListMessagesResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get chat messages
      tags:
      - messages
    post:
      consumes:
      - application/json
      description: Send a message to a chat
      parameters:
      - description: Chat ID
        in: path
        name: id
        required: true
        type: integer
      - description: Message text
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateMessageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CreateMessageResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Send a message
      tags:
      - messages
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"